+++ lib/filter-schema.js
@@ -7,9 +7,67 @@ const {
   TransformObjectFields
 } = require('@graphql-tools/wrap')
 
+const kDirectiveNamespace = Symbol('mercurius-auth.namespace')
+
 const { GraphQLObjectType } = require('graphql')
 
-module.exports.filterSchema = async function filter (graphQLSchema, policies, context) {
+module.exports = function filterIntrospectionSchema (app, policy, { namespace, applyPolicy: policyFunction }) {
+  if (!app[kDirectiveNamespace]) {
+    app[kDirectiveNamespace] = {}
+
+    // the filter hook must be the last one to be executed (after all the authContextHook ones)
+    app.ready(err => {
+      /* istanbul ignore next */
+      if (err) throw err
+      app.graphql.addHook('preExecution', filterGraphQLSchemaHook(namespace).bind(app))
+    })
+  }
+
+  if (app[kDirectiveNamespace][namespace]) {
+    app[kDirectiveNamespace][namespace].push({
+      policy,
+      policyFunction
+    })
+  } else {
+    app[kDirectiveNamespace][namespace] = [{
+      policy,
+      policyFunction
+    }]
+  }
+}
+
+function filterGraphQLSchemaHook (namespace) {
+  return async function filterHook (schema, document, context) {
+    if (!isIntrospection(document)) {
+      return
+    }
+
+    const filteredSchema = await filterSchema(schema,
+      this[kDirectiveNamespace][namespace],
+      context)
+
+    return {
+      schema: filteredSchema
+    }
+  }
+}
+
+function isIntrospection (document) {
+  // TODO switch the logic: exit when one non-introspection operation is found
+  const queryTypes = document.definitions.filter(def => def.operation === 'query')
+  for (const qt of queryTypes) {
+    // TODO: __Schema, __Type, __TypeKind, __Field, __InputValue, __EnumValue, __Directive
+    if (qt.selectionSet.selections.some(sel => (
+      sel.name.value === '__schema' ||
+      sel.name.value === '__type'
+    ))) {
+      return true
+    }
+  }
+  return false
+}
+
+async function filterSchema (graphQLSchema, policies, context) {
   const filterDirectiveMap = {}
 
   // each `policies` item is a directive
